backend/requirements.txt
absl-py==0.13.0
anaconda-client==1.7.2
anaconda-navigator==2.0.3
appdirs==1.4.4
argh==0.26.2
asgiref==3.4.1
astunparse==1.6.3
atomicwrites==1.4.0
backports.weakref==1.0.post1
bkcharts==0.2
black==19.10b0
boto==2.49.0
Bottleneck==1.3.2
brotlipy==0.7.0
cachetools==4.2.2
certifi==2020.12.5
clyent==1.2.2
comtypes==1.1.9
conda==4.10.3
conda-build==3.21.4
conda-verify==3.4.2
contextlib2==0.6.0.post1
cycler==0.10.0
cytoolz==0.11.0
Django==3.2.7
django-cors-headers==3.8.0
django-filter==2.4.0
djangorestframework==3.12.4
entrypoints==0.3
et-xmlfile==1.0.1
fastcache==1.1.0
flatbuffers==1.12
future==0.18.2
gast==0.4.0
google-auth==1.34.0
google-auth-oauthlib==0.4.5
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
HeapDict==1.0.1
jdcal==1.4.1
json5==0.9.5
jupyter==1.0.0
Keras==2.4.3
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
llvmlite==0.36.0
locket==0.2.1
Markdown==3.3.4
MarkupSafe==1.1.1
mccabe==0.6.1
menuinst==1.4.16
mistune==0.8.4
mkl-fft==1.3.0
mkl-service==2.3.0
mpmath==1.2.1
multipledispatch==0.6.0
mypy-extensions==0.4.3
navigator-updater==0.2.1
numpy==1.19.5
oauthlib==3.1.1
olefile==0.46
opt-einsum==3.3.0
parso==0.7.0
pathspec==0.7.0
patsy==0.5.1
pep8==1.7.1
pkginfo==1.7.0
ply==3.11
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
Pillow==8.2.0
pycurl==7.43.0.6
pyodbc===4.0.0-unsupported
pyreadline==2.1
pytest==6.2.3
pywin32==227
pywin32-ctypes==0.2.0
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==20.0.0
QDarkStyle==2.8.1
QtPy==1.9.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.18.1
simplegeneric==0.8.1
sip==4.19.13
sqlparse==0.4.1
statsmodels==0.12.2
tables==3.6.1
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
terminado==0.9.4
unicodecsv==0.14.1
webencodings==0.5.1
widgetsnbextension==3.5.1
win-unicode-console==0.5
wincertstore==0.2
wrapt==1.12.1
xlwings==0.23.0
xlwt==1.3.0
xmltodict==0.12.0
zict==2.0.0
zope.event==4.5.0
